arrows.js w/ lava arrows
This is a modified version of the arrows.js script that is included with the ScriptCraft plugin.  It adds a sixth arrow type, lava arrows, which does exactly what the name says.

The bits that were modified for this are line 40 (adding 'lava' to the list of arrow types), and lines 121-125, which implements the drone function at the location where the arrow lands in order to place a block of lava (10 is the block type id number for lava).